+++ discoveryapis_generator/lib/src/generated_googleapis/discovery/v1.dart
@@ -90,9 +90,9 @@ class ApisResource {
     };
 
     final _url = 'apis/' +
-        commons.Escaper.ecapeVariable('$api') +
+        commons.escapeVariable('$api') +
         '/' +
-        commons.Escaper.ecapeVariable('$version') +
+        commons.escapeVariable('$version') +
         '/rest';
 
     final _response = await _requester.request(